moved windows permission logic to new file
