+++ cli/user_perm_win.go
@@ -1,0 +1,29 @@
+package cli
+
+import (
+	"bytes"
+	"os/exec"
+)
+
+func elevatedRun(name string, arg ...string) (bool, error) {
+	ok, err := run("cmd", nil, append([]string{"/C", name}, arg...)...)
+	if err != nil {
+		ok, err = run("elevate.cmd", nil, append([]string{"cmd", "/C", name}, arg...)...)
+	}
+	return ok, err
+}
+
+func run(name string, dir *string, arg ...string) (bool, error) {
+	c := exec.Command(name, arg...)
+	if dir != nil {
+		c.Dir = *dir
+	}
+	var stderr bytes.Buffer
+	c.Stderr = &stderr
+	err := c.Run()
+	if err != nil {
+		return false, err
+	}
+
+	return true, nil
+}
